agregar desktop navbar
