+++ src/App.jsx
@@ -2,8 +2,9 @@ import CssBaseline from "@mui/material/CssBaseline";
 import { BrowserRouter, Routes, Route } from "react-router";
 
 import { CartPage } from "./modules/cart/pages/CartPage";
-import { Layout } from "./modules/core/components/Layout";
+import { Layout } from "./modules/core/layouts/Layout";
 import { AppRoutes } from "./modules/core/lib/AppRoutes";
+import { FavoritesPage } from "./modules/favorites/pages/FavoritesPage";
 import { HomePage } from "./modules/home/pages/HomePage";
 import { SearchPage } from "./modules/search/pages/SearchPage";
 
@@ -17,6 +18,7 @@ export default function App() {
             <Route index path={AppRoutes.homePage} element={<HomePage />} />
             <Route path={AppRoutes.searchPage} element={<SearchPage />} />
             <Route path={AppRoutes.cartPage} element={<CartPage />} />
+            <Route path={AppRoutes.favoritesPage} element={<FavoritesPage />} />
           </Route>
         </Routes>
       </BrowserRouter>
