+++ src/modules/core/components/DesktopNavbar.jsx
@@ -1,0 +1,26 @@
+import { CartButton } from "../../cart/components/CartButton";
+import { FavoritesButton } from "./FavoritesButton";
+import { Logo } from "./Logo";
+import { SearchForm } from "./SearchForm";
+
+export function DesktopNav() {
+  return (
+    <nav
+      style={{
+        display: "flex",
+        justifyContent: "space-between",
+        alignItems: "center",
+        gap: "1rem"
+      }}
+    >
+      <Logo />
+      <div style={{ flex: 1, maxWidth: "35rem" }}>
+        <SearchForm />
+      </div>
+      <div>
+        <CartButton />
+        <FavoritesButton />
+      </div>
+    </nav>
+  );
+}
