+++ src/modules/core/layouts/Header.jsx
@@ -1,0 +1,15 @@
+import { Container, useMediaQuery, useTheme } from "@mui/material";
+
+import { DesktopNav } from "../components/DesktopNavbar";
+import { MobileNav } from "../components/MobileNavbar";
+
+export function Header() {
+  const theme = useTheme();
+  const isMobileView = useMediaQuery(theme.breakpoints.down("sm"));
+
+  return (
+    <header>
+      <Container>{isMobileView ? <MobileNav /> : <DesktopNav />}</Container>
+    </header>
+  );
+}
