+++ src/modules/core/lib/AppRoutes.js
@@ -2,4 +2,5 @@ export class AppRoutes {
   static homePage = "/";
   static searchPage = "/buscar";
   static cartPage = "/carrito";
+  static favoritesPage = "/favoritos";
 }
