+++ src/modules/favorites/pages/FavoritesPage.jsx
@@ -1,0 +1,11 @@
+import { Container } from "@mui/material";
+
+export function FavoritesPage() {
+  return (
+    <>
+      <Container>
+        <h1>Favorites Page</h1>
+      </Container>
+    </>
+  );
+}
